Add DateHelper for setting/getting dates
Add a setDateFilter for filtering specials formats
