+++ galette/lib/Galette/Filters/ContributionsList.php
@@ -21,6 +21,7 @@
 
 namespace Galette\Filters;
 
+use Galette\Helpers\DatesHelper;
 use Throwable;
 use Analog\Analog;
 use Galette\Core\Pagination;
@@ -46,6 +47,8 @@ use Galette\Core\Pagination;
 
 class ContributionsList extends Pagination
 {
+    use DatesHelper;
+
     public const ORDERBY_DATE = 0;
     public const ORDERBY_BEGIN_DATE = 1;
     public const ORDERBY_END_DATE = 2;
@@ -154,33 +157,12 @@ class ContributionsList extends Pagination
                 switch ($name) {
                     case 'start_date_filter':
                     case 'end_date_filter':
-                        if ($this->$name === null) {
-                            return $this->$name;
-                        }
-                        try {
-                            $d = \DateTime::createFromFormat(__("Y-m-d"), $this->$name);
-                            if ($d === false) {
-                                //try with non localized date
-                                $d = \DateTime::createFromFormat("Y-m-d", $this->$name);
-                                if ($d === false) {
-                                    throw new \Exception('Incorrect format');
-                                }
-                            }
-                            return $d->format(__("Y-m-d"));
-                        } catch (Throwable $e) {
-                            //oops, we've got a bad date :/
-                            Analog::log(
-                                'Bad date (' . $this->$name . ') | ' .
-                                $e->getMessage(),
-                                Analog::INFO
-                            );
-                            return $this->$name;
-                        }
+                        return $this->getDate($name);
                     case 'rstart_date_filter':
                     case 'rend_date_filter':
                         //same as above, but raw format
                         $rname = substr($name, 1);
-                        return $this->$rname;
+                        return $this->getDate($rname, true, false);
                     default:
                         return $this->$name;
                 }
@@ -236,79 +218,7 @@ class ContributionsList extends Pagination
             switch ($name) {
                 case 'start_date_filter':
                 case 'end_date_filter':
-                    try {
-                        if ($value !== '') {
-                            $y = \DateTime::createFromFormat(__("Y"), $value);
-                            if ($y !== false) {
-                                $month = 1;
-                                $day = 1;
-                                if ($name === 'end_date_filter') {
-                                    $month = 12;
-                                    $day = 31;
-                                }
-                                $y->setDate(
-                                    (int)$y->format('Y'),
-                                    $month,
-                                    $day
-                                );
-                                $this->$name = $y->format('Y-m-d');
-                            }
-
-                            $ym = \DateTime::createFromFormat(__("Y-m"), $value);
-                            if ($y === false && $ym !== false) {
-                                $day = 1;
-                                if ($name === 'end_date_filter') {
-                                    $day = (int)$ym->format('t');
-                                }
-                                $ym->setDate(
-                                    (int)$ym->format('Y'),
-                                    (int)$ym->format('m'),
-                                    $day
-                                );
-                                $this->$name = $ym->format('Y-m-d');
-                            }
-
-                            $d = \DateTime::createFromFormat(__("Y-m-d"), $value);
-                            if ($y === false && $ym === false && $d !== false) {
-                                $this->$name = $d->format('Y-m-d');
-                            }
-
-                            if ($y === false && $ym === false && $d === false) {
-                                $formats = array(
-                                    __("Y"),
-                                    __("Y-m"),
-                                    __("Y-m-d"),
-                                );
-
-                                $field = null;
-                                if ($name === 'start_date_filter') {
-                                    $field = _T("start date filter");
-                                }
-                                if ($name === 'end_date_filter') {
-                                    $field = _T("end date filter");
-                                }
-
-                                throw new \Exception(
-                                    sprintf(
-                                        //TRANS: %1$s is field name, %2$s is list of known date formats
-                                        _T('Unknown date format for %1$s.<br/>Know formats are: %2$s'),
-                                        $field,
-                                        implode(', ', $formats)
-                                    )
-                                );
-                            }
-                        } else {
-                            $this->$name = null;
-                        }
-                    } catch (Throwable $e) {
-                        Analog::log(
-                            'Wrong date format. field: ' . $name .
-                            ', value: ' . $value . ', expected fmt: ' .
-                            __("Y-m-d") . ' | ' . $e->getMessage(),
-                            Analog::INFO
-                        );
-                        throw $e;
-                    }
+                    $this->setFilterDate($name, $value, $name === 'start_date_filter');
                     break;
                 default:
                     $this->$name = $value;
